# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Link Crew Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List discussion posts
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a discussion post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post with its comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post with comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{id}/comments:
    post:
      operationId: createComment
      tags: [posts]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{id}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [posts]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # FAQs
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [freshman, leader, admin]
        schoolId:
          type: integer
        schoolName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        privacyLevel:
          type: string
          enum: [public, school_only, private]
        quizCompleted:
          type: boolean
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, schoolId, schoolName, privacyLevel, quizCompleted, createdAt]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [classes, schedules, sat_act, clubs, sports, teachers, general, advice]
        commentCount:
          type: integer
        isPinned:
          type: boolean
        createdAt:
          type: string
      required: [id, schoolId, authorId, author, title, content, category, commentCount, isPinned, createdAt]
    PostDetail:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [classes, schedules, sat_act, clubs, sports, teachers, general, advice]
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        isPinned:
          type: boolean
        createdAt:
          type: string
      required: [id, schoolId, authorId, author, title, content, category, comments, isPinned, createdAt]
    PostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [classes, schedules, sat_act, clubs, sports, teachers, general, advice]
        schoolId:
          type: integer
      required: [title, content, category, schoolId]
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        createdAt:
          type: string
      required: [id, postId, authorId, author, content, createdAt]
    CommentInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
...